Move builtin reducers into their own header
